refactor(eval): move sweep execution into agent runner
